Add files via upload
This project is a calculator I made that will ask you for 3 numbers and once press calculate it will calculate the max, min, mean, median, and range. I did this using a bootstrap template, HTML coding, and a .css file as well as a javascript file I had to come up with to allow this process to happen.